feat: add slice to search page
